requirements.txt
requests==2.25.1
python-dotenv==0.19.2
MySQL-python==1.2.5
